Create a __main__.py file
