+++ stltovoxel/__init__.py
@@ -1,4 +1,4 @@
-from .main import convert_file, convert_files, convert_mesh, convert_meshes
+from .convert import convert_file, convert_files, convert_mesh, convert_meshes
 
 __all__ = [
     'convert_file',
